Implement a few hacky workarounds for `typing` usage.
This does four things:
- Ignores all imports of `typing`
- Replaces all functions calls which are named *exactly* `cast` with `$arg2` (`cast(1, 2)` becomes `2`)
  - Note: this does not ensure that `cast` should be `typing.cast`... It only ever checks if the call is exactly `cast`
- Removes all if statements which are *exactly* `if TYPE_CHECKING:`
- Flattens all if statements which are *exactly* `if not TYPE_CHECKING:`

Note for the last two: there's a very limited evaulator added, which can evaluate `not` and a literal variable with the exact name `TYPE_CHECKING`.

It will not change any other if statements - and it will not work for `if typing.TYPE_CHECKING` or any other variation which a fully complete one _should_ work with.