Implement a few hacky workarounds for `typing` usage.
This does two things:
- Ignores all imports of `typing`
- Removes all if statements which are *exactly* `if TYPE_CHECKING:`
- Flattens all if statements which are *exactly* `if not TYPE_CHECKING:`

Note for the last two: there's a very limited evaulator added, which can evaluate `not` and a literal variable with the exact name `TYPE_CHECKING`.

It will not change any other if statements - and it will not work for `if typing.TYPE_CHECKING` or any other variation which a fully complete one _should_ work with.