Simplify direct import x.y.z statements
Instead of producing 'var x = {}; __nest__(x, 'y.z', ...);', this will produce 'var x = {'y': {'z': ...}};'

The result will be faster 'import' statements, since the runtime no longer has to parse 'x.y.z' strings,
and no longer has to copy all attributes of the 'z' module to a new module.

This has a side effect of letting other modules change the exported variables from a module independently
of the module's own representation of the variable, but this already happened when changing a module-level
variable. The two representations (in __all__ and the module's var) were already de-synchronized, this just
allows for more modification of the __all__ one from other modules.