Make int() function match python's int() more fully.
Python's int() floors and will panic for non-int values. This makes
transcrypt's int behave similarly (deviating from float behavior - not
sure if this is what we want, but it might be?).

The one difference that's left is that JavaScript's parseInt and
thus this int() function accept float values and simply floors them
afterwards instead of throwing an error. int('4.6') returns 4 in
Transcrypt where it throws an error in Python.